Server/src/chatbot/requirements.txt
transformers==4.37.2
torch==2.1.2
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.5
faiss-cpu==1.7.4
python-dotenv==1.0.0
fastapi==0.105.0
uvicorn==0.25.0
jinja2==3.1.2
pymongo==4.6.1
pandas==2.1.4
numpy==1.26.2
pydantic==2.5.2 